Fix issue where key policy is not valid at update
Fix the issue where cleanup fails to give permission for CMS to delete the inactive KMS key.

This is important because sometimes the consumer IAM principal gets deleted in AWS, but the KMS
key policy stays around, in which case, Amazon replaces the princnipal ARN with the princnipal 'ID'
behind the scenes.

If that key policy containing a principal 'ID' is ever updated then the update will fail because
a princnipal 'ID' is not considered to be a valid AWS principal.