Feature/audit logging (#117)
Enable audit logging:

Custom logback appender that in conjunction with the AuditLogEventProcessor will log audit events in a flattened JSON formate that is AWS Athena optimized and roll the logs every 5 minutes.

S3LogUploaderService uploads the logs to S3 partitioned in a manor that is also optimized with Athena.every time they are rolled. This service is also registered as a shutdown hook and will tell the appender to roll the log and upload it to S3 during a shutdown event.

Refactored the way config is bootstrapped into ConfigService.